+++ client/src/features/works/AboutSubsection.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { useLocation } from 'react-router-dom'
+import { Col, Row } from 'react-bootstrap'
 
-import StyledDataRow from '../../styles/shared/DataRow'
 import StyledHr from '../../styles/shared/Hr'
 import ExpandableList from '../common/ExpandableList'
 import TextValue from '../common/TextValue'
@@ -9,21 +9,22 @@ import TextLabel from '../common/TextLabel'
 import { isObjectOrWork } from '../../lib/util/uri'
 import theme from '../../styles/theme'
 import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
-import RecordLink from '../common/RecordLink'
+
+import SubjectHeadingsList from './SubjectHeadingsList'
 
 interface ILinkData {
   content: Array<Array<string> | string>
+  id: string
   label?: string
   expandColumns?: boolean
-  id?: string
 }
 
 // Show an expandable list of links with a label in the left column
 const AboutSubsection: React.FC<ILinkData> = ({
   content,
+  id,
   label,
   expandColumns = false,
-  id = 'link-container',
 }) => {
   const [isMobile, setIsMobile] = useState<boolean>(
     window.innerWidth < theme.breakpoints.md,
@@ -32,40 +33,46 @@ const AboutSubsection: React.FC<ILinkData> = ({
 
   useResizeableWindow(setIsMobile)
 
-  const values = content.map((c: Array<string> | string) => {
-    return (
-      <React.Fragment>
-        {Array.isArray(c) ? (
-          c.map((link, ind: number) => (
-            <React.Fragment key={link}>
-              <RecordLink url={link} /> {ind !== c.length - 1 ? ' -- ' : ''}
-            </React.Fragment>
-          ))
-        ) : (
-          <RecordLink url={c} />
-        )}
-      </React.Fragment>
-    )
-  })
+  const values = content.map((c: Array<string> | string) => (
+    <SubjectHeadingsList subjectHeadings={c} />
+  ))
 
   if (content && content.length > 0) {
     return (
-      <StyledDataRow className="row" id={id} data-testid={id}>
-        {label !== undefined && (
-          <TextLabel label={label} className="col-md-3 col-sm-12" />
-        )}
-        <ExpandableList
-          className="col-md-9 col-sm-12"
-          length={20}
-          itemSpacing="double"
-        >
-          <TextValue values={values} className="col-12" itemSpacing="double" />
-        </ExpandableList>
+      <Row id={id} data-testid={`${id}-row`}>
+        <Col xl={3} lg={3} md={12} sm={12}>
+          {label !== undefined && <TextLabel label={label} />}
+        </Col>
+        <Col xl={9} lg={9} md={12} sm={12}>
+          <Row>
+            {!isMobile && (
+              <React.Fragment>
+                <Col xs={7}>
+                  <dt>Subjects</dt>
+                </Col>
+                <Col xs={5}>
+                  <dt>Matching Works</dt>
+                </Col>
+              </React.Fragment>
+            )}
+            <ExpandableList
+              className="col-sm-12"
+              length={20}
+              itemSpacing="double"
+            >
+              <TextValue
+                values={values}
+                className="col-12"
+                itemSpacing="double"
+              />
+            </ExpandableList>
+          </Row>
+        </Col>
         <StyledHr
           className="linkContainerHr"
           hidden={expandColumns || (isObjectOrWork(pathname) && isMobile)}
         />
-      </StyledDataRow>
+      </Row>
     )
   }